feat: add API mocking, update components, add types
API mocking for development
To not overload the server with requests, we are using MSW (Mock Service Worker) to mock API calls. There were also issues with the live API, so we are using a local mock server for development.
Just to note, only products endpoint is mocked.

Updated Component
Create own components for the layout, Product listing and product detail. Also added skeleton loading for the product card and product detail.

Added types
Derived types for the product structure.

Misc
- Formated code with Prettier
- Added hero section to the home page